+++ packages/core-js/internals/check-iterator-closing-on-early-error.js
@@ -1,0 +1,11 @@
+'use strict';
+module.exports = function (exec, arg) {
+  var CLOSED = false;
+  try {
+    exec.call({
+      next: function () { return { done: true }; },
+      'return': function () { CLOSED = true; }
+    }, arg);
+  } catch (error) { /** empty */ }
+  return CLOSED;
+};